+++ examples/simple_debug.py
@@ -36,7 +36,7 @@ if __name__ == '__main__':
         SupRB(
             rule_generation=es.ES1xLambda(
                 operator='&',
-                init=rule.initialization.NormalInit(sigma=np.array([0.1,1]), fitness=rule.fitness.VolumeWu(alpha=0.8)),
+                init=rule.initialization.NormalInit(sigma=np.array([0.00000001,1]), fitness=rule.fitness.VolumeWu(alpha=0.8)),
                 mutation=suprb.optimizer.rule.mutation.Uniform(sigma=np.array([1,1]))
             ),
             solution_composition=ga.GeneticAlgorithm(
