+++ suprb/rule/initialization.py
@@ -156,12 +156,14 @@ class NormalInit(RuleInit):
 
     def gaussian_kernel_function(self, mean: np.ndarray, random_state: RandomState) -> MatchingFunction:
         center = random_state.normal(loc=mean, scale=self.sigma[0], size=(mean.shape[0]))
+        test3 = center - mean
+        #center = mean
+        t = self.sigma[0]
         # radius = interval ]0,1]
         radius = (1 - np.random.random_sample(size=(mean.shape[0])))
 
 
-        test = GaussianKernelFunction(center=center,
-                                      radius=radius)
+        test = GaussianKernelFunction(center=center,radius=radius)
 
         return test
