+++ tests/test_matching_functions.py
@@ -19,22 +19,6 @@ import itertools
 
 class TestMatchingFunction(unittest.TestCase):
     """Test Matching Function Implementation based on Higdon Gramacy Lee"""
-
-    def create_rule(self, fitness, experience, error):
-        center = np.array([0, 0])
-        radius = np.array([0.5, 0.1])
-
-        rule = Rule(match=GaussianKernelFunction(center, radius),
-                    input_space=[-1.0, 1.0],
-                    model=SupRB(),
-                    fitness=VolumeWu)
-
-        rule.fitness_ = fitness
-        rule.experience_ = experience
-        rule.error_ = error
-
-        return rule
-
     def setup_test_example(self):
 
         n_samples = 1000
@@ -99,72 +83,6 @@ class TestMatchingFunction(unittest.TestCase):
 
         return super().setUp()
 
-    def test_ellipsoid_volume(self):
-        # circle
-        ellipsoid = GaussianKernelFunction(center=np.array([0, 0]), radius=np.array([1, 1]))
-        self.assertAlmostEqual(ellipsoid.volume_, 3.1416, 4)
-        # sphere
-        ellipsoid = GaussianKernelFunction(center=np.array([0, 0, 0]), radius=np.array([1, 1, 1]))
-        self.assertAlmostEqual(ellipsoid.volume_, (4 / 3) * 3.1416, 4)
-        # sphere with different radius
-        ellipsoid = GaussianKernelFunction(center=np.array([0, 0, 0]), radius=np.array([0.5, 0.5, 0.5]))
-        self.assertAlmostEqual(ellipsoid.volume_, (4 / 3) * 3.1416 * 0.125, 4)
-
-    def test_ellipsoid_matching(self):
-        # circle -> are the center points inside the circle
-        points = np.array([[0, 0], [0, 0], [0, 0]])
-        ellipsoid = GaussianKernelFunction(center=np.array([0, 0]), radius=np.array([1, 1]))
-        self.assertTrue(all(ellipsoid.__call__(points)))
-
-        # circle -> are the points on the outer radius inside the circle
-        points = np.array([[0, 0.999], [0.999, 0]])
-        ellipsoid = GaussianKernelFunction(center=np.array([0, 0]), radius=np.array([1, 1]))
-        self.assertTrue(all(ellipsoid.__call__(points)))
-
-        # circle -> are the points on the outer radius inside the circle
-        points = np.array([[0.2, 0.3], [-0.1, -0.1]])
-        ellipsoid = GaussianKernelFunction(center=np.array([0.5, 0.5]), radius=np.array([0.4, 0.4]))
-        self.assertTrue(ellipsoid.__call__(points)[0])
-        self.assertFalse(ellipsoid.__call__(points)[1])
-
-        # sphere -> are the center points inside the sphere
-        points = np.array([[0, 0, 0]])
-        ellipsoid = GaussianKernelFunction(center=np.array([0, 0, 0]), radius=np.array([1, 1, 1]))
-        self.assertTrue(all(ellipsoid.__call__(points)))
-
-        # ellipsoid -> are the points in the ellipsoid
-        points = np.array([[0, 0, 0], [0.999, 0, 0], [0, 0.499, 0], [0, 0, -0.999]])
-        ellipsoid = GaussianKernelFunction(center=np.array([0, 0, 0]), radius=np.array([1, 0.5, -1]))
-        self.assertTrue(all(ellipsoid.__call__(points)))
-
-    def test_ellipsoid_clip(self):
-        center = np.array([0, 0.5, -0.7])
-        radius = np.array([8, 0, 7])
-
-        ellipsoid = GaussianKernelFunction(center=center, radius=radius)
-        ellipsoid.clip(np.array([]))
-
-        print(f"center: {ellipsoid.center} radius: {ellipsoid.radius}")
-
-        self.assertTrue(all(np.equal(ellipsoid.center, np.array([0, 0.5, -0.7]))))
-        np.testing.assert_almost_equal(ellipsoid.radius, np.array([1., 0.5, 0.3]), decimal=5)
-
-    def test_ellipsoid_mutate(self):
-        center = np.array([0, 0])
-
-        X = np.array([center - 0.001])
-
-        rule = self.create_rule(fitness=1, experience=1,error=0.01)
-        test = rule.match(X)
-        ruleMutate = Uniform(matching_type=GaussianKernelFunction, sigma=0.1)
-        ruleMutate.gaussian_kernel_function(rule=rule, random_state=42)
-        test1 = rule.match(X)
-
-        #
-
-        self.assertTrue(True)
-
-    @unittest.skip("isn't working on python 3.9")
     def test_smoke_test(self):
         for matching_func, mutation, initialization in self.combined_matching_params:
             self.setup_model(matching_func, mutation, initialization)
